More Wind Meter Changes
Grouped meter and added scale and translation to make it not overlap the label.
Increased pointer transparency to make it easier to read numbers when they're
obscured by the the pointer.
Simplified arc and tick drawing to remove magic numbers and reduce the number
of path elements for the ticks by half.
Added i18n to get a nicer label for the meter.