Handle new sources reactively
Just push all updates to streambundle and when new sources appear
react on newSources.onValue in IP: offer the new source to all the
existing widgets, allowing each to 'consume' the new source. If
nobody has consumed the source create a new widget for it by
path or if there is no special widget just create a new Digital.