fix: unit handling
Unit handling was broken because of changes in keyswithmetadata.json.
The fix offloads unit handling to signalk-schema.getUnits(path), which
at this stage is the place for it, because all schema awareness is there.
At a later stage this should probably be in signalk-client, which may
choose to query the server for its version or fall back to an embedded
schema.